+++ cmp/util.h
@@ -13,6 +13,6 @@ extern char tmp_buffer[];
 
 char *repeat(char c, size_t n);
 
-size_t count_following(unsigned int org, char *str, char c);
+size_t count_following(size_t org, char *str, char c);
 
 #endif //BFCMP_UTIL_H
